refactor(backend): extract card review owner
