+++ apps/ios/Flashcards/Flashcards/Database/CardStore/CardStore+FSRSRepair.swift
@@ -68,7 +68,7 @@ extension CardStore {
     }
 }
 
-// Keep in sync with apps/backend/src/cards/fsrs.ts::getInvalidFsrsStateReason.
+// Keep in sync with apps/backend/src/cards/review/fsrs.ts::getInvalidFsrsStateReason.
 func invalidFsrsStateReason(card: Card) -> String? {
     if card.fsrsCardState == .new {
         if card.dueAt != nil {
@@ -104,7 +104,7 @@ func invalidFsrsStateReason(card: Card) -> String? {
     return nil
 }
 
-// Keep in sync with apps/backend/src/cards/fsrs.ts repair semantics.
+// Keep in sync with apps/backend/src/cards/review/fsrs.ts repair semantics.
 func resetFsrsState(card: Card, updatedAt: String) -> Card {
     Card(
         cardId: card.cardId,
