+++ apps/ios/Flashcards/Flashcards/Database/LocalDatabase/LocalDatabase.swift
@@ -7,7 +7,7 @@ import Foundation
 
  This file mirrors the backend scheduler-settings and review-persistence logic
  in `apps/backend/src/scheduling/workspaceSettings.ts` and
- `apps/backend/src/cards/reviews.ts`.
+ `apps/backend/src/cards/review/reviews.ts`.
  If you change scheduler-state validation or review persistence here, make the
  same change in the backend mirror and update docs/fsrs-scheduling-logic.md.
